nose apparently requires runTest
