+++ mdsobjects/python/tests/Makefile.in
@@ -656,10 +656,6 @@ TESTS = \
         threadsUnitTest.py
 
 
-#        test_trees.py \
-#        test_data.py \
-#
-
 # DISABLE PYTHON LEAK DETECTION: This suppression file shall be removed when
 # python will finally introduce the patch to detect if valgrind is running ..
 # take a look at the file for details. You can also disable this if you are
